refactor: split presentation.tsx into focused component modules
Extract PreviewCard, ProgressDisplay, CompletionArtifact, ToolCallChip,
and messages into individual files. Move formatCurrency/formatRelativeTime
to shared utils. Presentation.tsx becomes a temporary barrel re-exporting
all names for backward compatibility (removal target: 2026-03-14).
